+++ source/common/upstream/subset_lb.h
@@ -38,8 +38,9 @@ private:
   // Represents a subset of an original HostSet.
   class HostSubsetImpl : public HostSetImpl {
   public:
-    HostSubsetImpl(const HostSet& original_host_set)
-        : HostSetImpl(original_host_set.priority()), original_host_set_(original_host_set) {}
+    HostSubsetImpl(const HostSet& original_host_set, bool locality_weight_aware)
+        : HostSetImpl(original_host_set.priority()), original_host_set_(original_host_set),
+          locality_weight_aware_(locality_weight_aware) {}
 
     void update(const HostVector& hosts_added, const HostVector& hosts_removed,
                 HostPredicate predicate);
@@ -49,12 +50,14 @@ private:
 
   private:
     const HostSet& original_host_set_;
+    const bool locality_weight_aware_;
   };
 
   // Represents a subset of an original PrioritySet.
   class PrioritySubsetImpl : public PrioritySetImpl {
   public:
-    PrioritySubsetImpl(const SubsetLoadBalancer& subset_lb, HostPredicate predicate);
+    PrioritySubsetImpl(const SubsetLoadBalancer& subset_lb, HostPredicate predicate,
+                       bool locality_weight_aware);
 
     void update(uint32_t priority, const HostVector& hosts_added, const HostVector& hosts_removed);
 
@@ -81,6 +84,7 @@ private:
   private:
     const PrioritySet& original_priority_set_;
     const HostPredicate predicate_;
+    const bool locality_weight_aware_;
     bool empty_ = true;
   };
 
@@ -155,6 +159,8 @@ private:
 
   // Forms a trie-like structure. Requires lexically sorted Host and Route metadata.
   LbSubsetMap subsets_;
+
+  const bool locality_weight_aware_;
 
   friend class SubsetLoadBalancerDescribeMetadataTester;
 };